Create a generic detect method
There was a lot of duplication in your detect methods.  Addresses was
different enough that I left it alone, but the rest were reduced to a
single method to keep things DRY.

Also, a set does not need to be turned back in to a list to iterate
through it.